Moved auth file to libs folder and added set reminder button
